pyproject.toml
[project]
name = "handtokening"
version = "0.4.0"
description = "Code-signing server built in Python/Django"
readme = "README.md"
authors = [
    { name = "Dexter Castor Döpping", email = "dexter.c.dopping@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "asn1crypto>=1.5.1",
    "django>=5.2.5",
    "django-ipware>=7.0.1",
    "djangorestframework>=3.16.1",
    "vt-py>=0.21.0",
]

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.9",
]

[tool.ruff.lint.per-file-ignores]
"**/handtokening/settings/*.py" = ["F403"]
